requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
pinecone-client==5.0.1
openai==1.93.0
python-dotenv==1.0.1
tiktoken==0.8.0
gunicorn==23.0.0
tqdm==4.66.5
beautifulsoup4==4.12.3
requests==2.32.3
Flask-Limiter==3.12
weaviate-client==4.5.4
nltk==3.9.1
timeout-decorator==0.5.0